+++ src/OpenSleigh.Persistence.Cosmos.SQL/Entities/SagaState.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace OpenSleigh.Persistence.Cosmos.SQL.Entities
 {
@@ -9,18 +7,6 @@ namespace OpenSleigh.Persistence.Cosmos.SQL.Entities
         public byte[] Data { get; set; } = null;
         public Guid? LockId { get; set; } = null;
         public DateTime? LockTime { get; set; } = null;
-    }
-
-    internal class SagaStateEntityTypeConfiguration : IEntityTypeConfiguration<SagaState>
-    {
-        public void Configure(EntityTypeBuilder<SagaState> builder)
-        {
-            builder.ToContainer("SagaStates")
-                            .HasNoDiscriminator();
-
-            builder.HasKey(e => new {e.CorrelationId, e.Type});
-            builder.HasPartitionKey(e => e.PartitionKey);
-        }
     }
 
 }