+++ src/OpenSleigh.Transport.Kafka/HeaderNames.cs
@@ -1,0 +1,8 @@
+﻿namespace OpenSleigh.Transport.Kafka
+{
+    public static class HeaderNames
+    {
+        public const string MessageType = "message-type";
+        public const string Error = "error";
+    }
+}