+++ src/OpenSleigh.Transport.Kafka/IQueueReferenceFactory.cs
@@ -1,0 +1,9 @@
+﻿using OpenSleigh.Core.Messaging;
+
+namespace OpenSleigh.Transport.Kafka
+{
+    public interface IQueueReferenceFactory
+    {
+        QueueReferences Create<TM>(TM message = default) where TM : IMessage;
+    }
+}